Move AbstractMspMoteType into MspMoteType
All MSP430 motes extend AbstractMspMoteType,
so move the methods in that class into the
MspMoteType class instead. This will allow
removing AbstractMspMoteType.

The reason to move AbstractMspMoteType into
MspMoteType and not the other way around
is to align the MSP mote functionality
with the Contiki mote functionality.
