+++ java/org/contikios/cooja/mspmote/AbstractMspMoteType.java
@@ -34,150 +34,13 @@
  */
 
 package org.contikios.cooja.mspmote;
-import java.awt.Container;
-import java.awt.Image;
-import java.io.File;
-import java.net.URL;
-
-import javax.swing.Icon;
-import javax.swing.ImageIcon;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
 
 import org.contikios.cooja.MoteInterface;
-import org.contikios.cooja.MoteType;
-import org.contikios.cooja.Simulation;
-import org.contikios.cooja.dialogs.CompileContiki;
-import org.contikios.cooja.dialogs.MessageContainer;
-import org.contikios.cooja.dialogs.MessageList;
 
 /**
  *
  */
 public abstract class AbstractMspMoteType extends MspMoteType {
-
-    protected final Logger logger = LogManager.getLogger(getClass());
-
-    public abstract String getMoteType();
-
-    public abstract String getMoteName();
-
-    protected abstract String getMoteImage();
-
-    @Override
-    public boolean configureAndInit(Container parentContainer, Simulation simulation, boolean visAvailable)
-            throws MoteTypeCreationException {
-        /* If visualized, show compile dialog and let user configure */
-        if (visAvailable && !simulation.isQuickSetup()) {
-
-            /* Create unique identifier */
-            if (getIdentifier() == null) {
-                int counter = 0;
-                boolean identifierOK = false;
-                while (!identifierOK) {
-                    identifierOK = true;
-
-                    counter++;
-                    setIdentifier(getMoteType() + counter);
-
-                    for (MoteType existingMoteType : simulation.getMoteTypes()) {
-                        if (existingMoteType == this) {
-                            continue;
-                        }
-                        if (existingMoteType.getIdentifier().equals(getIdentifier())) {
-                            identifierOK = false;
-                            break;
-                        }
-                    }
-                }
-            }
-
-            /* Create initial description */
-            if (getDescription() == null) {
-                setDescription(getMoteName() + " Mote Type #" + getIdentifier());
-            }
-
-            return MspCompileDialog.showDialog(parentContainer, simulation, this, getMoteType());
-        }
-
-        /* Not visualized: Compile Contiki immediately */
-        if (getIdentifier() == null) {
-            throw new MoteTypeCreationException("No identifier");
-        }
-
-        final MessageList compilationOutput = MessageContainer.createMessageList(visAvailable);
-
-        if (getCompileCommands() != null) {
-            /* Handle multiple compilation commands one by one */
-            String[] arr = getCompileCommands().split("\n");
-            for (String cmd: arr) {
-                cmd = cmd.trim();
-                if (cmd.isEmpty()) {
-                    continue;
-                }
-
-                try {
-                    CompileContiki.compile(
-                            cmd,
-                            null,
-                            getContikiSourceFile().getParentFile(),
-                            null,
-                            null,
-                            compilationOutput,
-                            true
-                            );
-                } catch (Exception e) {
-                    /* Print last 10 compilation errors to console */
-                    MessageContainer[] messages = compilationOutput.getMessages();
-                    for (int i = Math.max(messages.length - 10, 0); i < messages.length; i++) {
-                        logger.fatal(">> " + messages[i]);
-                    }
-
-                    logger.fatal("Compilation error: " + compilationOutput);
-                    return false;
-                }
-            }
-        }
-
-        if (getContikiFirmwareFile() == null
-                || !getContikiFirmwareFile().exists()) {
-            throw new MoteTypeCreationException(
-                    "Contiki firmware file does not exist: "
-                            + getContikiFirmwareFile());
-        }
-        return true;
-    }
-
-    @Override
-    public Icon getMoteTypeIcon() {
-        String imageName = getMoteImage();
-        if (imageName == null) {
-            return null;
-        }
-        URL imageURL = this.getClass().getClassLoader().getResource(imageName);
-        if (imageURL == null) {
-            return null;
-        }
-        ImageIcon icon = new ImageIcon(imageURL);
-        if (icon.getIconHeight() > 0 && icon.getIconWidth() > 0) {
-            Image image = icon.getImage().getScaledInstance(
-                    (200 * icon.getIconWidth() / icon.getIconHeight()), 200,
-                    Image.SCALE_DEFAULT);
-            return new ImageIcon(image);
-        }
-        return null;
-    }
-
-    @Override
-    public File getExpectedFirmwareFile(File source) {
-        File parentDir = source.getParentFile();
-        String sourceNoExtension = source.getName();
-        if (sourceNoExtension.endsWith(".c")) {
-            sourceNoExtension = sourceNoExtension.substring(0, source.getName().length() - 2);
-        }
-        return new File(parentDir, "/build/" + getMoteType() + "/" + sourceNoExtension + '.' + getMoteType());
-    }
-
     protected Class<? extends MoteInterface>[] createMoteInterfaceList(Class<? extends MoteInterface>... interfaceList) {
         return interfaceList;
     }
